logging updates and optimize some code

+++ facade/service.go
@@ -1257,38 +1257,38 @@ func (f *Facade) fillServiceAddr(ctx datastore.Context, svc *service.Service) er
 
 	for idx := range svc.Endpoints {
 		endpointName := svc.Endpoints[idx].Name
-		//TODO: only lookup if there is a possibility for an address assignment. i.e. AddressConfig has port and protocol
-		if assignment, err := f.FindAssignmentByServiceEndpoint(ctx, svc.ID, endpointName); err != nil {
-			glog.Errorf("Error searching for address assignments for endpoint %s of service %s (%s): %s", endpointName, svc.Name, svc.ID, err)
-			return err
-		} else if assignment != nil {
-			// verify the ports match
-			if port := svc.Endpoints[idx].AddressConfig.Port; assignment.Port != port {
-				glog.Infof("Removing address assignment for endpoint %s of service %s (%s)", endpointName, svc.Name, svc.ID)
-				if err := store.Delete(ctx, addressassignment.Key(assignment.ID)); err != nil {
-					glog.Errorf("Error removing address assignment for endpoint %s of service %s (%s): %s", endpointName, svc.Name, svc.ID, err)
-					return err
-				}
-				svc.Endpoints[idx].RemoveAssignment()
-				continue
-			}
-
-			// verify the ip exists
-			if exists, err := f.HasIP(ctx, svc.PoolID, assignment.IPAddr); err != nil {
-				glog.Errorf("Error validating address assignment for endpoint %s of service %s (%s): %s", endpointName, svc.Name, svc.ID, err)
+		//make sure there is no assignment; shouldn't be but just in case
+		svc.Endpoints[idx].RemoveAssignment()
+		//only lookup if there is a possibility for an address assignment. i.e. AddressConfig has port and protocol
+		if svc.Endpoints[idx].IsConfigurable() {
+			if assignment, err := f.FindAssignmentByServiceEndpoint(ctx, svc.ID, endpointName); err != nil {
+				glog.Errorf("Error searching for address assignments for endpoint %s of service %s (%s): %s", endpointName, svc.Name, svc.ID, err)
 				return err
-			} else if !exists {
-				glog.Infof("Removing address assignment for endpoint %s of service %s (%s)", endpointName, svc.Name, svc.ID)
-				if err := store.Delete(ctx, addressassignment.Key(assignment.ID)); err != nil {
-					glog.Errorf("Error removing address assignment for endpoint %s of service %s (%s): %s", endpointName, svc.Name, svc.ID, err)
-					return err
+			} else if assignment != nil {
+				// verify the ports match
+				if port := svc.Endpoints[idx].AddressConfig.Port; assignment.Port != port {
+					glog.Infof("Removing address assignment for endpoint %s of service %s (%s)", endpointName, svc.Name, svc.ID)
+					if err := store.Delete(ctx, addressassignment.Key(assignment.ID)); err != nil {
+						glog.Errorf("Error removing address assignment for endpoint %s of service %s (%s): %s", endpointName, svc.Name, svc.ID, err)
+						return err
+					}
+					continue
 				}
-				svc.Endpoints[idx].RemoveAssignment()
-				continue
+
+				// verify the ip exists
+				if exists, err := f.HasIP(ctx, svc.PoolID, assignment.IPAddr); err != nil {
+					glog.Errorf("Error validating address assignment for endpoint %s of service %s (%s): %s", endpointName, svc.Name, svc.ID, err)
+					return err
+				} else if !exists {
+					glog.Infof("Removing address assignment for endpoint %s of service %s (%s)", endpointName, svc.Name, svc.ID)
+					if err := store.Delete(ctx, addressassignment.Key(assignment.ID)); err != nil {
+						glog.Errorf("Error removing address assignment for endpoint %s of service %s (%s): %s", endpointName, svc.Name, svc.ID, err)
+						return err
+					}
+					continue
+				}
+				svc.Endpoints[idx].SetAssignment(*assignment)
 			}
-			svc.Endpoints[idx].SetAssignment(*assignment)
-		} else {
-			svc.Endpoints[idx].RemoveAssignment()
 		}
 	}
 	return nil
